Add files via upload
Inspired by the Smart Beta strategy, this quantitative trading strategy utilized risk appetite to construct model. This strategy aims to capture and take advantage of the differences in market risk attitudes changes of various stocks. When the market risk attitude is highly aggressive, actively invest in a postive appetite stock portfolio in pursuit of high returns; when the market risk attitude is extremely conservative, actively invest in a negative appetite portfolio. The strategy pursues safety, so as to achieve a relative balance between returns and risks.